requirement.txt
sqlalchemy
sqlalchemy[asyncio] 
asyncpg
psycopg2-binary
python-dotenv
alembic
bcrypt